Move launch fragment to fragment package
